Added function to read in dictionary
Made the program able to read in a dictionary by a defined local file,
and dynamically allocate an array for the dictionary. Also, created a
test function to determine if the dictionary is being read in or not.

TODO: Change where Clion compiles the executible for the sake of being
able to read in the dictionary without having to copy files constantly
between directories